+++ src/pages/LoanPage/CreditCard/CreditCard.module.css
@@ -1,0 +1,64 @@
+.cardPromo {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 40px 60px;
+  background: linear-gradient(to bottom, #d8e4f0, #fee0b8);
+  box-shadow:
+    0px 2px 4px 0px #00000033,
+    0px 0px 8px 0px #00000014;
+  border-radius: 28px;
+  margin: 40px auto;
+}
+
+.cardPromo__title {
+  font-size: 48px;
+  color: #ffffff;
+  margin-bottom: 16px;
+}
+
+.cardPromo__description {
+  font-weight: 500;
+  font-size: 16px;
+  letter-spacing: 2%;
+  color: #2d3748;
+  max-width: 515px;
+}
+
+.cardPromo__conditionsValue {
+  color: #3d3d3d;
+  font-weight: 700;
+  font-size: 20px;
+  letter-spacing: 2%;
+  margin-bottom: 10px;
+}
+
+.cardPromo__conditionsLabel {
+  color: #3d3d3d;
+  font-weight: 500;
+  font-size: 16px;
+  letter-spacing: 2%;
+}
+
+.cardPromo__textColumn {
+  display: flex;
+  flex-direction: column;
+  gap: 35px;
+}
+
+.cardPromo__conditionsItem {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
+
+.cardPromo__conditions {
+  display: flex;
+  gap: 30px;
+  flex-wrap: wrap;
+}
+
+.cardPromo__button {
+  max-width: 145px;
+  border-radius: 8px;
+}
