+++ src/pages/LoanPage/LoanPage.module.css
@@ -1,39 +1,3 @@
-.cardPromo {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 40px 60px;
-  background: linear-gradient(to bottom, #d8e4f0, #fee0b8);
-  box-shadow:
-    0px 2px 4px 0px #00000033,
-    0px 0px 8px 0px #00000014;
-  border-radius: 28px;
-  margin: 40px auto;
-}
-
-.cardPromo__conditions {
-  display: flex;
-  gap: 30px; /* Отступ между колонками характеристик */
-  flex-wrap: wrap; /* Перенос на новую строку, если не хватает места */
-}
-
-.cardPromo__conditionsItem {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start; /* Выравнивание по левому краю внутри элемента */
-}
-
-.cardPromo__conditionsValue {
-  font-size: 1.4em;
-  font-weight: 600;
-  color: #003cff; /* Акцентный цвет (из твоих предыдущих задач) */
-}
-
-.cardPromo__conditionsLabel {
-  font-size: 0.9em;
-  color: #777;
-}
-
 /* --- Общий контейнер страницы --- */
 /* --- Общий контейнер страницы --- */
 /* --- Общий контейнер страницы --- */
@@ -50,20 +14,20 @@
 /* --- Секция CreditCardHero --- */
 .creditCardHero {
   display: flex;
-  align-items: center; /* Центрируем контент по вертикали */
-  justify-content: space-between; /* Распределяем колонки по краям */
-  gap: 60px; /* Отступ между колонками */
-  margin-bottom: 80px; /* Отступ между Hero-секцией и табами */
-  background-color: #e6f7ff; /* Легкий фон для Hero */
+  align-items: center;
+  justify-content: space-between;
+  gap: 60px;
+  margin-bottom: 80px;
+  background-color: #e6f7ff;
   padding: 40px;
   border-radius: 12px;
 }
 
 .creditCardHero__textColumn {
-  flex: 1; /* Позволяет колонке занимать доступное пространство */
+  flex: 1;
   display: flex;
   flex-direction: column;
-  align-items: flex-start; /* Выравнивание по левому краю */
+  align-items: flex-start;
   text-align: left;
 }
 
@@ -82,14 +46,13 @@
 }
 
 .creditCardHero__conditions {
-  /* Стили для условий, например, если это список */
   margin-bottom: 30px;
   font-size: 1em;
   color: #666;
 }
 
 .creditCardHero__conditions ul {
-  list-style: disc; /* Или 'none' если используете свои иконки */
+  list-style: disc;
   margin-left: 20px;
   padding: 0;
 }
@@ -99,13 +62,12 @@
 }
 
 .creditCardHero__button {
-  /* Если нужно переопределить стили кнопки для этой секции */
   padding: 15px 30px;
   font-size: 1.1em;
 }
 
 .creditCardHero__imageColumn {
-  flex: 1; /* Позволяет колонке занимать доступное пространство */
+  flex: 1;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -120,8 +82,8 @@
 /* --- Секция табов (cardTabs) --- */
 .cardTabs__navigation {
   display: flex;
-  justify-content: flex-start; /* Табы выравниваем по левому краю */
-  border-bottom: 1px solid #e0e0e0; /* Общая тонкая линия под табами */
+  justify-content: flex-start;
+  border-bottom: 1px solid #e0e0e0;
   margin-bottom: 30px;
   position: relative; /* Для жирной линии */
 }
@@ -129,14 +91,14 @@
 .cardTabs__button {
   background: none;
   border: none;
-  padding: 15px 25px; /* Отступы для кнопок табов */
+  padding: 15px 25px;
   font-size: 1.1em;
   font-weight: 600;
   color: #777;
   cursor: pointer;
   transition: color 0.3s ease;
-  position: relative; /* Важно для ::after */
-  outline: none; /* Убираем стандартный outline при фокусе, можно добавить кастомный */
+  position: relative;
+  outline: none;
 }
 
 .cardTabs__button:hover {
@@ -152,62 +114,17 @@
   content: "";
   position: absolute;
   left: 0;
-  bottom: -1px; /* -1px, чтобы перекрыть тонкую линию border-bottom у родителя */
+  bottom: -1px;
   width: 100%;
-  height: 3px; /* Толщина жирной линии */
-  background-color: #007bff; /* Цвет жирной линии */
+  height: 3px;
+  background-color: #007bff;
   border-radius: 2px 2px 0 0;
 }
 
 .cardTabs__content {
-  padding: 20px 0; /* Отступ между навигацией и содержимым таба */
-  /* Здесь можно добавить стили для контейнера содержимого, например, border */
+  padding: 20px 0;
 }
 
 /* --- Медиа-запросы для адаптивности --- */
 @media (max-width: 992px) {
-  .creditCardHero {
-    flex-direction: column; /* Колонки Hero-секции располагаются в столбец */
-    gap: 40px;
-    padding: 30px;
-    text-align: center; /* Центрируем текст */
-  }
-
-  .creditCardHero__textColumn {
-    align-items: center; /* Центрируем элементы внутри текстовой колонки */
-  }
-
-  .creditCardHero__title {
-    font-size: 2.2em;
-  }
-
-  .creditCardHero__description {
-    font-size: 1em;
-  }
-
-  .cardTabs__navigation {
-    flex-wrap: wrap; /* Табы будут переноситься на новую строку */
-    justify-content: center; /* Центрируем табы */
-  }
-
-  .cardTabs__button {
-    padding: 10px 15px;
-    font-size: 1em;
-  }
-}
-
-@media (max-width: 576px) {
-  .loanPage {
-    padding: 20px 10px;
-  }
-  .creditCardHero {
-    padding: 20px;
-  }
-  .creditCardHero__title {
-    font-size: 1.8em;
-  }
-  .creditCardHero__button {
-    padding: 12px 20px;
-    font-size: 1em;
-  }
 }
